Bootstraps control plane and node
Includes tests for single node and a scenario for cluster. The cluster
scenario test does work in that you DO get a two node cluster created
with Molecule. But, the k8s cluster falls apart quickly as far as the
deployments becoming unstable and failing once the second node joins in.